hf_spaces_deployment/requirements.txt
# Core AI and ML dependencies
groq>=0.4.0
python-dotenv>=1.0.0
gradio>=4.44.0

# Audio processing 
speechrecognition>=3.10.0
pydub>=0.25.0
# Note: PyAudio not needed for browser-based recording via Gradio

# Text-to-speech
gtts>=2.3.0
elevenlabs>=0.2.0

# Image processing
Pillow>=9.0.0
pillow-avif-plugin>=0.3.0

# Additional dependencies for Hugging Face Spaces
requests>=2.28.0
numpy>=1.21.0

# Web audio processing
ffmpeg-python>=0.2.0